Organização e separação do código (criação de headers)
