+++ README.md
@@ -1,2 +1,8 @@
 # EDII-T1
 Estrutura de Dados II - Trabalho 1 
+
+Tarefas para o próximo commit:
+- ˜Organização do código.˜
+- Interface de Usuário.
+- Verificar como deve ser feito o dump (aguardando e-mail).
+- Deletar registro (fragmentação)
